Refactor PreviewData
