+++ CycleSnap/Configurations/RealmConfiguration.swift
@@ -29,38 +29,3 @@ class RealmConfiguration {
         _ = try! Realm()
     }
 }
-
-extension Realm {
-    static var previewRealm: Realm {
-        let identifier = "previewRealm"
-        let config = Realm.Configuration(inMemoryIdentifier: identifier)
-        do {
-            let realm = try Realm(configuration: config)
-
-            let categoryList = realm.objects(Category.self)
-            if categoryList.isEmpty {
-                try realm.write {
-                    for i in 0 ... 2 {
-                        // Creating sample category
-                        let category = Category()
-                        category.name = "SampleCategory\(i)"
-                        category.orderIndex = i
-                        category.isLatestFirst = true
-
-                        for j in 0 ... 4 {
-                            // Creating sample photo
-                            let photo = Photo()
-                            photo.path = "photos/" + category.name + "/sample\(j).jpg"
-                            category.photos.append(photo)
-                        }
-
-                        realm.add(category)
-                    }
-                }
-            }
-            return realm
-        } catch {
-            fatalError("Can't bootstrap item data: \(error.localizedDescription)")
-        }
-    }
-}
